Add level scripts

+++ Assets/Scripts/Models/DTO/HourlyData.cs
@@ -5,9 +5,9 @@ public class HourlyData
     [JsonProperty("time")]
     public string[] Time { get; set; }
     [JsonProperty("temperature_2m")]
-    public double[] Temperature { get; set; }
+    public float[] Temperature { get; set; }
     [JsonProperty("windspeed_10m")]
-    public double[] WindSpeed { get; set; }
+    public float[] WindSpeed { get; set; }
     [JsonProperty("winddirection_10m")]
     public int[] WindDirection { get; set; }
 }
